Hopefully fix build issue with Coord typedef.
The build failed on MacOS Sonoma and Visual Studio 17 2022.

The problem was that we were redefining the min, max, round and abs
functions for Coord.  I don't think that is necessary, because
Coord is just a typedef for float and the standard functions from
<cmath> should work for that.

I wasn't able to test this locally, so we'll have to look at the
feedback from GitHub Actions to see if this fix worked or not.